refactor(plugin-api): split monolithic lib.rs into modular file structure
Break up the 473-line lib.rs into organized modules: host.rs (HostApiC), vtable.rs (PluginVTable/PluginInstanceC), types/ (metadata, content, context, theme, shortcut, i18n). Re-exports maintained for backward compatibility. Update plugin-dev.md to match new import paths.
